Se cambia tipo de datos para la latitud y longitud, se modifica el servicio del crear y responder comentarios, se modifica el metodo de validar coordenadas, se modifica servicio de recuperar password

+++ src/main/java/co/edu/uniquindio/proyecto/modelo/Ubicacion.java
@@ -10,6 +10,6 @@ import lombok.*;
 @EqualsAndHashCode
 public class Ubicacion {
 
-    private int latitud;
-    private int longitud;
+    private Double latitud;
+    private Double longitud;
 }
